+++ src/main/java/cecs429/weight/Strategy.java
@@ -3,7 +3,7 @@ package cecs429.weight;
 import cecs429.index.DiskPositionalIndex;
 
 public interface Strategy {
-    public Float getWdt(Integer tftd, Integer tokensInDoc, double aveTokensInCorpus, double aveTftd);
+    public Float getWdt(Integer tftd, double tokensInDoc, double aveTokensInCorpus, double aveTftd);
     public Float getWqt(Integer corpusSize, Integer dft);
-    public Float getLd(double docWeight, Integer byteSize);
+    public Float getLd(double docWeight, double byteSize);
 }
