+++ src/main/java/cecs429/weight/Wacky.java
@@ -4,7 +4,7 @@ import cecs429.index.DiskPositionalIndex;
 
 public class Wacky implements Strategy{
     @Override
-    public Float getWdt(Integer tftd, Integer tokensInDoc, double aveTokensInCorpus, double aveTftd) {
+    public Float getWdt(Integer tftd, double tokensInDoc, double aveTokensInCorpus, double aveTftd) {
         return (float)((1+Math.log(tftd)/Math.log(Math.E))/(1+Math.log(aveTftd)/Math.log(Math.E)));
     }
 
@@ -14,7 +14,7 @@ public class Wacky implements Strategy{
     }
 
     @Override
-    public Float getLd(double docWeight, Integer byteSize) {
+    public Float getLd(double docWeight, double byteSize) {
         return (float)Math.sqrt(byteSize);
     }
